Initial content
This has 3 primary files:
1. `transactions.nim` to download the transaction from the stack account
2. `totals.sh` to extract the monthly totals using `jq`
3. `types.nim` which defines the basic data structure of each JSON

By default, the transactions are saved in a file `transactions.json`
which isn't valid json. It needs a comma at the end of every line except
for the last an open bracked at the beginning and a close bracked at the
end to make it a valid JSON array.

The `config.nims` is to allow enums with the same names and enforce
NEP-1 style